+++ packages/koot-boilerplate-legacy/src/views/static/index.jsx
@@ -1,22 +1,6 @@
 import React from 'react'
 import { extend } from 'koot'
 
-const PageStaticAssets = ({
-    className
-}) =>
-    <div className={className}>
-        <div className="block">
-            <h2>{__('pages.static.method_require')}</h2>
-            <p>{__('pages.static.method_require_content')}</p>
-            <img src={require('@assets/photo.jpg')} />
-        </div>
-        <div className="block">
-            <h2>{__('pages.static.method_static')}</h2>
-            <p>{__('pages.static.method_static_content')}</p>
-            <img src="/photo.jpg" />
-        </div>
-    </div>
-
 export default extend({
     pageinfo: (/*state, renderProps*/) => ({
         title: `${__('pages.static.title')} - ${__('title')}`,
@@ -26,4 +10,20 @@ export default extend({
         ]
     }),
     styles: require('./styles.less'),
-})(PageStaticAssets)
+})(
+    ({
+        className
+    }) =>
+        <div className={className}>
+            <div className="block">
+                <h2>{__('pages.static.method_require')}</h2>
+                <p>{__('pages.static.method_require_content')}</p>
+                <img src={require('@assets/photo.jpg')} />
+            </div>
+            <div className="block">
+                <h2>{__('pages.static.method_static')}</h2>
+                <p>{__('pages.static.method_static_content')}</p>
+                <img src={(__SPA__ ? '' : '/') + "photo.jpg"} />
+            </div>
+        </div>
+)